Refactor : 출결표 테이블 컴포넌트 분리.
