+++ src/Pages/MainPage/index.jsx
@@ -1,19 +1,14 @@
 import { useState, useEffect, useContext } from 'react';
 
 import { AuthContext } from 'App';
-import { checkCloumns, validDay, isValidCheck } from 'Utils';
+import { validDay } from 'Utils';
 import AllTableService from 'Network/AllTableService';
 
 import { format } from 'date-fns';
 
 import { CusDatePicker, ShowToday } from 'Components';
-import PopoverCheckAttend from './PopoverCheckAttend';
-import WrongDay from './WrongDay';
 import UserGuide from './UserGuide';
-import Tabs from '@mui/material/Tabs';
-import Tab from '@mui/material/Tab';
-import Box from '@mui/material/Box';
-import { DataGrid } from '@mui/x-data-grid';
+import MainPageTable from './MainPageTable';
 
 import Styled from './MainPage.styled';
 
@@ -23,9 +18,6 @@ const MainPage = () => {
   const [tab, setTab] = useState(0);
   const [rowData, setRowData] = useState(null);
   const [selectRowData, setSelectRowData] = useState(null);
-
-  const [anchorEl, setAnchorEl] = useState(null);
-  const [curFocus, setCurFocus] = useState({ id: '', select: '' });
 
   const auth = useContext(AuthContext);
   const userId = auth.status.userId;
@@ -47,73 +39,6 @@ const MainPage = () => {
   const handleChangeTab = (event, dstTab) => {
     setTab(dstTab);
     updateSelectRowData(rowData, dstTab);
-  };
-
-  const handleClickCell = (params, event) => {
-    const field = params.field;
-    if (field !== 'checkIn' && field !== 'checkOut') return;
-
-    const selectUserInfo = selectRowData.find(array => array.id === params.id);
-    const myInfo = rowData.find(array => array.id === +userId);
-    // console.log('myinfo :', myInfo);
-    // console.log('selectUserInfo : ', selectUserInfo);
-    if (myInfo === undefined) {
-      alert('이번 달에 참가하고 있지 않습니다!');
-      return;
-    }
-
-    const valid = isValidCheck(
-      selectUserInfo,
-      myInfo.id,
-      myInfo.role,
-      myInfo.team,
-    );
-    // console.log('valud:', valid);
-    if (valid) {
-      valid === -1 ? alert('수정 권한이 없습니다!') : alert('다른 팀입니다!');
-      return;
-    }
-    setAnchorEl(event.currentTarget);
-    setCurFocus({ id: params.id, select: field });
-  };
-
-  const handleChangeCheck = async value => {
-    const id = curFocus.id;
-    const select = curFocus.select;
-    const today = new Date();
-    let result;
-    // console.log('id : ', id, value);
-    // console.log('row', selectRowData);
-    const selectUserInfo = selectRowData.find(array => array.id === id);
-
-    if (value === 6 && selectUserInfo.vacation === 0) {
-      alert('사용할 수 있는 휴가가 없습니다!');
-      setAnchorEl(null);
-      return;
-    }
-    if (
-      today.getFullYear() !== date.getFullYear() ||
-      today.getMonth() !== date.getMonth()
-    ) {
-      alert('지난 달 기록은 수정할 수 없습니다!');
-      setAnchorEl(null);
-      return;
-    }
-    if (select === 'checkIn') {
-      result = await AllTableService.putAllTableCheckIn({
-        userId: id,
-        checkIn: value,
-        tableDay: format(date, 'yyyy-MM-dd'),
-      });
-    } else {
-      result = await AllTableService.putAllTableCheckOut({
-        userId: id,
-        checkOut: value,
-        tableDay: format(date, 'yyyy-MM-dd'),
-      });
-    }
-    setAnchorEl(null);
-    getUsers();
   };
 
   const getUsers = async () => {
@@ -162,32 +87,14 @@ const MainPage = () => {
         <CusDatePicker date={date} setDate={setDate} filterWeekend={true} />
       </div>
       <Styled.MainTable>
-        <Box className="table">
-          <Tabs value={tab} onChange={handleChangeTab}>
-            <Tab label="전체 보기" />
-            <Tab label="레드 팀" />
-            <Tab label="블루 팀" />
-          </Tabs>
-          {validDay(date) ? (
-            <WrongDay wrongType={validDay(date)} />
-          ) : (
-            <>
-              <DataGrid
-                rows={selectRowData}
-                columns={checkCloumns}
-                onCellClick={handleClickCell}
-                hideFooterPagination={true} // 페이지 네이션 비활성화, 전체, 빨간팀, 파란팀?
-                hideFooterSelectedRowCount={true} // row count 숨기기
-                getRowClassName="cell"
-              />
-              <PopoverCheckAttend
-                anchorEl={anchorEl}
-                setAnchorEl={setAnchorEl}
-                onChangeCheck={handleChangeCheck}
-              />
-            </>
-          )}
-        </Box>
+        <MainPageTable
+          tab={tab}
+          handleChangeTab={handleChangeTab}
+          date={date}
+          rowData={rowData}
+          selectRowData={selectRowData}
+          userId={userId}
+        />
       </Styled.MainTable>
     </Styled.MainBackground>
   );
